fix(spring-core):处理代码风格问题
