--- spring-core/spring-context-util_test.go
@@ -1,36 +1,0 @@
-/*
- * Copyright 2012-2019 the original author or authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package SpringCore
-
-import (
-	"github.com/go-spring/go-spring/spring-core/testdata/bar"
-	"github.com/go-spring/go-spring/spring-core/testdata/foo"
-	"github.com/stretchr/testify/assert"
-	"testing"
-)
-
-func TestGetBeanUname(t *testing.T) {
-	beanUname := getBeanUnameByType(checkBeanType(&foo.Demo{}))
-	otherPkgBeanUname := getBeanUnameByType(checkBeanType(&bar.Demo{}))
-	assert.NotEqual(t, beanUname, otherPkgBeanUname)
-}
-
-func TestMustPointerTypeOf(t *testing.T) {
-	assert.Panics(t, func() {
-		checkBeanType(foo.Demo{})
-	})
-}
